+++ src/main/java/net/silentchaos512/gear/event/ServerEvents.java
@@ -33,7 +33,7 @@ public final class ServerEvents {
 
         // Send crafting items packets to correct for registry changes
         SilentGear.LOGGER.debug("Sending materials craftin item correction packet");
-        Network.channel.sendTo(new SyncMaterialCraftingItemsPacket(), playerMP.connection.netManager, NetworkDirection.PLAY_TO_CLIENT);
+        Network.channel.sendTo(new SyncMaterialCraftingItemsPacket(MaterialManager.getValues()), playerMP.connection.netManager, NetworkDirection.PLAY_TO_CLIENT);
         SilentGear.LOGGER.debug("Sending parts crafting item correction packet");
         Network.channel.sendTo(new SyncGearCraftingItemsPacket(), playerMP.connection.netManager, NetworkDirection.PLAY_TO_CLIENT);